requirements.txt

aiogram==3.4.1
python-dotenv==1.0.0

aiosqlite==0.19.0
asyncpg==0.29.0


requests==2.31.0
pandas>=2.2.0
numpy==1.26.2


matplotlib==3.8.2
Pillow==10.1.0

transformers==4.36.0
torch==2.1.2
huggingface-hub==0.20.1
tokenizers==0.15.0
sentencepiece==0.1.99
accelerate==0.25.0

textblob==0.17.1
vaderSentiment==3.3.2

beautifulsoup4==4.12.2
lxml==4.9.3

scikit-learn==1.3.2
scipy==1.11.4

